requirements.txt
streamlit
gitpython
pinecone-client
groq
sentence-transformers
PyGithub
langchain-pinecone
langchain-community